using size_t when it comes to integer to pointer conversions

+++ main/polyobj.c
@@ -11,101 +11,8 @@ AND AGREES TO THE TERMS HEREIN AND ACCEPTS THE SAME BY USE OF THIS FILE.
 COPYRIGHT 1993-1998 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 */
 /*
- * $Source: /cvsroot/dxx-rebirth/d1x-rebirth/main/polyobj.c,v $
- * $Revision: 1.1.1.1 $
- * $Author: zicodxx $
- * $Date: 2006/03/17 19:43:43 $
  * 
  * Hacked-in polygon objects
- * 
- * $Log: polyobj.c,v $
- * Revision 1.1.1.1  2006/03/17 19:43:43  zicodxx
- * initial import
- *
- * Revision 1.2  2002/03/26 08:20:01  donut
- * fix briefing spinning robot display in OGL
- *
- * Revision 1.1.1.1  1999/06/14 22:11:08  donut
- * Import of d1x 1.37 source.
- *
- * Revision 2.1  1995/05/26  16:10:37  john
- * Support for new 4-byte align v8 pof files.
- * 
- * Revision 2.0  1995/02/27  11:32:44  john
- * New version 2.0, which has no anonymous unions, builds with
- * Watcom 10.0, and doesn't require parsing BITMAPS.TBL.
- * 
- * Revision 1.64  1995/01/14  19:16:43  john
- * First version of new bitmap paging code.
- * 
- * Revision 1.63  1994/12/14  18:06:54  matt
- * Removed compile warnings
- * 
- * Revision 1.62  1994/12/09  17:54:31  john
- * Made the CFILE's close right after reading in data.
- * 
- * Revision 1.61  1994/12/09  16:13:28  mike
- * speedup pof file reading, but still horribly slow using hog file...problem somewhere else.
- * 
- * Revision 1.60  1994/12/08  17:41:20  yuan
- * Cfiling stuff.
- * 
- * Revision 1.59  1994/11/21  11:02:19  matt
- * Added error checking
- * 
- * Revision 1.58  1994/11/14  11:32:49  matt
- * Allow switching to simpler models even when alt_textures specified
- * 
- * Revision 1.57  1994/11/13  21:15:24  matt
- * Added basic support for more than one level of detail simplification
- * 
- * Revision 1.56  1994/11/11  19:29:25  matt
- * Added code to show low detail polygon models
- * 
- * Revision 1.55  1994/11/10  14:02:57  matt
- * Hacked in support for player ships with different textures
- * 
- * Revision 1.54  1994/11/03  11:01:59  matt
- * Made robot pics lighted
- * 
- * Revision 1.53  1994/11/02  16:18:34  matt
- * Moved draw_model_picture() out of editor
- * 
- * Revision 1.52  1994/10/18  14:38:11  matt
- * Restored assert now that bug is fixed
- * 
- * Revision 1.51  1994/10/17  21:35:03  matt
- * Added support for new Control Center/Main Reactor
- * 
- * Revision 1.50  1994/10/14  17:46:23  yuan
- * Made the soft Int3 only work in net mode.
- * 
- * Revision 1.49  1994/10/14  17:43:47  yuan
- * Added soft int3's instead of Asserts  for some common network bugs.
- * 
- * Revision 1.48  1994/10/14  17:09:04  yuan
- * Made Assert on line 610 be if in an attempt
- * to bypass.
- * 
- * Revision 1.47  1994/09/09  14:23:42  matt
- * Added glow code to polygon models for engine glow
- * 
- * Revision 1.46  1994/08/26  18:03:30  matt
- * Added code to remap polygon model numbers by matching filenames
- * 
- * Revision 1.45  1994/08/26  15:35:58  matt
- * Made eclips usable on more than one object at a time
- * 
- * Revision 1.44  1994/08/25  18:11:58  matt
- * Made player's weapons and flares fire from the positions on the 3d model.
- * Also added support for quad lasers.
- * 
- * Revision 1.43  1994/07/25  00:14:18  matt
- * Made a couple of minor changes for the drivethrough
- * 
- * Revision 1.42  1994/07/25  00:02:41  matt
- * Various changes to accomodate new 3d, which no longer takes point numbers
- * as parms, and now only takes pointers to points.
  * 
  */
 
@@ -904,7 +811,7 @@ extern int polymodel_read_n(polymodel *pm, int n, CFILE *fp)
 	for (i = 0; i < n; i++) {
 		pm[i].n_models = read_int(fp);
 		pm[i].model_data_size = read_int(fp);
-		pm[i].model_data = (ubyte *) read_int(fp);
+		pm[i].model_data = (ubyte *) (size_t)read_int(fp);
 		for (j = 0; j < MAX_SUBMODELS; j++)
 			pm[i].submodel_ptrs[j] = read_int(fp);
 		for (j = 0; j < MAX_SUBMODELS; j++)
